pyproject.toml
[project]
name = "image-classification-ml2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastai>=2.8.2",
    "kaggle>=1.7.4.5",
    "pillow>=11.2.1",
    "pip>=25.1.1",
    "timm>=1.0.15",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]
